Separate core from hubot. WIP
